+++ CHANGELOG.md
@@ -23,6 +23,7 @@ Each release is divided into the following main categories:
   import 'allink-core-static/js/modules/expandable';
   ```
 ### NEW
+- Site Overlay: New custom events added. Details in comments of `js/modules/site-overlay.js`.
 - Google Maps: Markers can now be opened with custom link. Details in comments of `js/modules/map.js`.
 - Google Maps: `FullscreenControl` is now set to `false` per default. This can be overwriiten by setting the variable `fullscreen_enabled` to `true` on a project basis.
 - Lazyload loader icon is now removed from the DOM when image is loaded (with a delay to make the icon is hidden after the image animation has finished)
@@ -41,4 +42,4 @@ Each release is divided into the following main categories:
 - Infobox: Added new plugin
 - docCookies, Added module
 ### FIXES
--
+- Site Overlay: In case `click-close-enabled` is active, the cursor is now correctly set to pointer.